requirements.txt
asgiref==3.8.1
bootstrap4==0.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
et-xmlfile==1.1.0
idna==3.8
openpyxl==3.1.5
psycopg2-binary==2.9.9
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2